Revert "Sell view"
