--- clearoutt.xcodeproj/project.xcworkspace/xcuserdata/bolanleadisa.xcuserdatad/WorkspaceSettings.xcsettings
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>BuildLocationStyle</key>
-	<string>UseAppPreferences</string>
-	<key>CustomBuildLocationType</key>
-	<string>RelativeToDerivedData</string>
-	<key>DerivedDataLocationStyle</key>
-	<string>Default</string>
-	<key>ShowSharedSchemesAutomaticallyEnabled</key>
-	<true/>
-</dict>
-</plist>
