+++ clearoutt/CameraView.swift
@@ -1,0 +1,19 @@
+//
+//  CameraView.swift
+//  clearoutt
+//
+//  Created by Bolanle Adisa on 1/22/24.
+//
+
+import SwiftUI
+
+struct CameraView: View {
+    var body: some View {
+        Text("Camera Functionality Here")
+        // Implement the camera functionality using UIImagePickerController and SwiftUI integration
+    }
+}
+
+#Preview {
+    CameraView()
+}
