--- clearoutt/College.swift
@@ -1,13 +1,0 @@
-//
-//  College.swift
-//  clearoutt
-//
-//  Created by Bolanle Adisa on 1/25/24.
-//
-
-import Foundation
-
-struct College: Identifiable, Decodable, Hashable {
-    var id: String { name }
-    let name: String
-}
