--- clearoutt/GoogleService-Info.plist
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>API_KEY</key>
-	<string>AIzaSyAM58ClazlWyLomWPhqxaXaqx_oaWFC_pY</string>
-	<key>GCM_SENDER_ID</key>
-	<string>799444167082</string>
-	<key>PLIST_VERSION</key>
-	<string>1</string>
-	<key>BUNDLE_ID</key>
-	<string>com.cocoa.clearout</string>
-	<key>PROJECT_ID</key>
-	<string>clearout-82860</string>
-	<key>STORAGE_BUCKET</key>
-	<string>clearout-82860.appspot.com</string>
-	<key>IS_ADS_ENABLED</key>
-	<false></false>
-	<key>IS_ANALYTICS_ENABLED</key>
-	<false></false>
-	<key>IS_APPINVITE_ENABLED</key>
-	<true></true>
-	<key>IS_GCM_ENABLED</key>
-	<true></true>
-	<key>IS_SIGNIN_ENABLED</key>
-	<true></true>
-	<key>GOOGLE_APP_ID</key>
-	<string>1:799444167082:ios:c8eca9c08f7b888e78e3d3</string>
-</dict>
-</plist>
