--- clearoutt/RentView.swift
@@ -1,18 +1,0 @@
-//
-//  RentView.swift
-//  clearoutt
-//
-//  Created by Bolanle Adisa on 1/24/24.
-//
-
-import SwiftUI
-
-struct RentView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    RentView()
-}
