+++ clearoutt/ShopView.swift
@@ -1,18 +1,18 @@
 //
-//  DonateView.swift
+//  ShopView.swift
 //  clearoutt
 //
-//  Created by Bolanle Adisa on 1/24/24.
+//  Created by Bolanle Adisa on 1/22/24.
 //
 
 import SwiftUI
 
-struct DonateView: View {
+struct ShopView: View {
     var body: some View {
         Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
     }
 }
 
 #Preview {
-    DonateView()
+    ShopView()
 }
