+++ clearoutt/UserSession.swift
@@ -13,13 +13,7 @@ class UserSession: ObservableObject {
     // Add properties here to represent the user's session
     // For example, a simple isLoggedIn flag:
     @Published var isLoggedIn: Bool = false
-    @Published var isAuthenticated = false
-    @Published var firstName: String = ""
 
     // Add more properties and methods as needed
-    
-    func authenticateUser() {
-            isAuthenticated = true
-        }
 }
