+++ clearoutt/clearouttApp.swift
@@ -6,20 +6,12 @@
 //
 
 import SwiftUI
-import Firebase
 
 @main
 struct clearouttApp: App {
-    
-    init() {
-        FirebaseApp.configure()
-    }
-    
     var body: some Scene {
         WindowGroup {
             ContentView()
-            .environmentObject(UserSession())
         }
     }
 }
-
